move : errors dir -> exception dir
