+++ app/src/main/java/com/codesoom/assignment/exception/DecodingInValidTokenException.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.errors;
+package com.codesoom.assignment.exception;
 
 public class DecodingInValidTokenException extends RuntimeException {
 	public DecodingInValidTokenException(String token) {
